% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{aggregate_interval}
\alias{aggregate_interval}
\title{Aggregation by interval}
\usage{
aggregate_interval(depth, start_date, end_date, frequency, project = NULL)
}
\arguments{
\item{depth}{numeric. Defines the depth of the VWC values to aggregate.}

\item{start_date}{character. Defines the start date of the aggregation period.}

\item{end_date}{character. Defines the end date of the aggregation period.}

\item{frequency}{numeric. Defines the number of days of the aggregation time interval.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
None
}
\description{
Aggregates VWC values according to starting date & time interval and computes the basic statistics (mean, min, max, sd) of each time interval.
}
\examples{
\dontrun{ aggregate_interval(5, "2020-06-15", "2020-08-15", 7, "WI_region") }
}
