% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{aggregate_sum}
\alias{aggregate_sum}
\title{Aggregation by dates}
\usage{
aggregate_sum(depth, start_date, end_date, project = NULL)
}
\arguments{
\item{depth}{numeric. Defines the depth of the VWC values to visualize.}

\item{start_date}{character. Defines the start date of the aggregation period.}

\item{end_date}{character. Defines the end date of the aggregation period.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
None
}
\description{
Summarizes the basic statistics (mean, min, max, sd) of VWC values within one time interval.
}
\examples{
\dontrun{ aggregate_sum(5, "2020-06-15", "2020-08-15", project="WI_region") }
}
