% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{area_sum}
\alias{area_sum}
\title{Temporal variation for the entire area}
\usage{
area_sum(depth, project = NULL)
}
\arguments{
\item{depth}{numeric. Defines the depth of the VWC values to visualize.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
a time series plot visualizing the basic statistics of the region.
}
\description{
Visualizes the change in basic statistics (mean, min, median, max) of the regional VWC values across time.
}
\examples{
\dontrun{ area_sum(5, project="WI_region") }
}
