% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_region_prediction.R
\name{download_map}
\alias{download_map}
\title{Sub-region map downloading and processing}
\usage{
download_map(
  start_date,
  end_date,
  resolution,
  project = NULL,
  sub_area = NULL,
  filename = "sub_regions.shp"
)
}
\arguments{
\item{start_date}{character. Defines the start date of the prediction period. The start date should be no earlier than 2016-01-01.}

\item{end_date}{character. Defines the end date of the prediction period.}

\item{resolution}{numeric. Defines the resolution of mapping region. The resolution should be smaller than or equal to 500m.}

\item{project}{character. Defines the name of the folder to be created and save the maps. If not defined, the maps will be saved in the working directory.}

\item{sub_area}{numeric or vector of two or more numeric values. Defines the sub regions to call the function on.}

\item{filename}{character. Defines the file name in which the split map is stored (in shp format). Default as sub_regions.shp.}
}
\value{
None
}
\description{
Downloads and process the maps of each sub region. The output will be saved in a folder
under the name of the index of the sub region.
}
\examples{
\dontrun{ download_map("2021-04-01", "2021-11-01", 30, "Grant2021", sub_area=1:2) }
}
