% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{model_rz}
\alias{model_rz}
\title{ML model for rootzone VWC prediction based on existing data}
\format{
A list with 23 objects
}
\usage{
model_rz
}
\description{
A dataset containing the soil root zone ML model based on the existing
VWC data obtained from sites all around the US.
}
\keyword{datasets}
