% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{model_surface}
\alias{model_surface}
\title{ML model for surface VWC prediction}
\format{
A list with 23 objects
}
\usage{
model_surface
}
\description{
A dataset containing the soil surface level ML model based on the existing
VWC data obtained from sites all around the US.
}
\keyword{datasets}
