% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_region_prediction.R
\name{mosaic_region}
\alias{mosaic_region}
\title{Mosaic sub regions}
\usage{
mosaic_region(
  start_date,
  end_date,
  resolution,
  percentile = FALSE,
  project = NULL,
  sub_area = NULL,
  filename = "sub_regions.shp"
)
}
\arguments{
\item{start_date}{character. Defines the start date of the prediction period. The start date should be no earlier than 2016-01-01.}

\item{end_date}{character. Defines the end date of the prediction period.}

\item{resolution}{numeric. Defines the resolution of mapping region. The resolution should be smaller than or equal to 500m.}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will also be mapped.}

\item{project}{character. Defines the name of the folder to be created and save the maps. If not defined, the maps will be saved in the working directory.}

\item{filename}{character. Defines the file name in which the split map is stored (in shp format). Default as sub_regions.shp.}
}
\value{
None
}
\description{
Predicts the VWC values of each sub regions separately, using mean of covariates across all
sub regions to fill the NA values. The function then merges the downloaded maps of sub regions
back into a large-region map as the final output of the predicted VWC of the large region.
}
\examples{
\dontrun{ mosaic_region("2020-06-01", "2020-06-03", 500,T,"large_region") }
}
