% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{pixelwise_sum}
\alias{pixelwise_sum}
\title{Pixel-wise variation over time}
\usage{
pixelwise_sum(depth, project = NULL)
}
\arguments{
\item{depth}{numeric. Defines the depth of the VWC values to visualize.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
a leaflet plot visualizing the basic statistics (mean, min, median, max) of the pixel-wise variation in the region.
}
\description{
Visualizes the pixel-wise variations in the regional VWC values across time.
NOTE: very time consuming for maps with high resolution and large area.
}
\examples{
\dontrun{ pixelwise_sum(5, project="WI_region") }
}
