% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{plot_aggregated_VWC}
\alias{plot_aggregated_VWC}
\title{Aggregation visualization}
\usage{
plot_aggregated_VWC(
  depth,
  start_date,
  end_date,
  date = NULL,
  frequency = NULL,
  project = NULL
)
}
\arguments{
\item{depth}{numeric. Defines the depth of the VWC values to visualize.}

\item{start_date}{character. Defines the start date of the aggregation period.}

\item{end_date}{character. Defines the end date of an aggregation period saved by aggregation_by_date().}

\item{date}{character. Defines the starting date of an interval saved by aggregation_summary().}

\item{frequency}{numeric. Defines the length of the aggregation time interval (in days).}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
a leaflet map with layers of the basic statistics of the interval.
}
\description{
Visualizes the output maps saved by the two aggregation functions.
}
\examples{
\dontrun{ plot_aggregated_VWC(5, "2020-06-15", "2020-08-15", date="2020-06-29", frequency=14, project="WI_region")
          plot_aggregated_VWC(5, "2020-06-15", "2020-08-15", project="WI_region") }
}
