% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{plot_map}
\alias{plot_map}
\title{Daily map visualization}
\usage{
plot_map(date, depth, percentile = FALSE, project = NULL)
}
\arguments{
\item{date}{character. Defines the date of the visualized maps}

\item{depth}{numeric. Defines the depth of the VWC values to visualize.}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will be included.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}
}
\value{
a leaflet map containing the VWC maps.
}
\description{
Visualizes the maps created by the regional prediction function (one day, one depth at a time).
}
\examples{
\dontrun{ plot_map("2020-06-15", 5, TRUE, project="WI_region") }
}
