% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{point_CI}
\alias{point_CI}
\title{Validation time series}
\usage{
point_CI(
  test,
  depth,
  project = NULL,
  site_ = NULL,
  filename = "VWC_point_data.csv"
)
}
\arguments{
\item{test}{character. Defines the file name of the validation (measured) data (in csv format).}

\item{depth}{numeric. Indicating the depth of the predicted and measured data.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}

\item{site_}{character or a vector of characters. Defines the specific points to include in the plot. If not defined, all points will be plotted (for site number less than or equal to 12). If site number exceeds 12, 12 points will be chosen at random.}

\item{filename}{character. Defines the file name where the predicted data is stored (in csv format). Default as VWC_point_data.csv.}
}
\value{
a time series plot with predicted and measured data, as well as the 90 percent CI of the predicted data.
}
\description{
Compares the predicted and measured data via time series plot with 90 percent Confidence Interval.
}
\examples{
\dontrun{ point_CI("Huges_VWC.csv", 5, project="WI_point") }
}
