% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_plot.R
\name{point_extraction}
\alias{point_extraction}
\title{Point Extraction}
\usage{
point_extraction(
  site_,
  percentile = FALSE,
  project = NULL,
  filename = "VWC_point_data.csv"
)
}
\arguments{
\item{site_}{character. Defines the file name that stores the location of the points (in shapefile format).}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will be included.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}

\item{filename}{character. Defines the file name where the predicted data is stored (in csv format). Default as VWC_point_data.csv.}

\item{depth}{numeric. Indicating the depth of the map to extract.}
}
\value{
None
}
\description{
Extracts VWC values of specified points from the predicted regional maps.
The point map should be saved within project folder for function access.
}
\examples{
\dontrun{ point_extraction("WI_SM.shp", TRUE, project="WI_region") }
}
