% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{point_performance}
\alias{point_performance}
\title{Validation}
\usage{
point_performance(
  test,
  depth,
  stats = FALSE,
  project = NULL,
  filename = "VWC_point_data.csv"
)
}
\arguments{
\item{test}{character. Defines the file name of the validation (measured) data (in csv format).}

\item{depth}{numeric. Indicating the depth of the predicted and measured data.}

\item{stats}{logical. If TRUE, the values of r-squared, RMSE, bias and KGE will be included in the diagnosis plot.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}

\item{filename}{character. Defines the file name where the predicted data is stored (in csv format). Default as VWC_point_data.csv.}
}
\value{
a list of validation values including r-squared, RMSE, NSE, KGE, bias, RPD, and RPIQ.
}
\description{
Plots the predicted v. measured scatter plot.
}
\examples{
\dontrun{ point_performance("Huges_VWC.csv", 5, TRUE, project="WI_point") }
}
