% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_plot.R
\name{site_variation}
\alias{site_variation}
\title{Point Variation}
\usage{
site_variation(
  depth,
  percentile = FALSE,
  project = NULL,
  site_ = NULL,
  filename = "VWC_point_data.csv"
)
}
\arguments{
\item{depth}{numeric. Indicating the depth of the plotted VWC.}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will be included.}

\item{project}{character. Defines the name of the folder where the maps are saved. If not defined, the function will look for maps in the working directory.}

\item{site_}{character or a vector of characters. Defines the specific points to include in the plot. If not defined, all points will be plotted (for site number less than or equal to 12). If site number exceeds 12, 12 points will be chosen at random.}

\item{filename}{character. Defines the file name where the predicted data is stored (in csv format). Default as VWC_point_data.csv.}
}
\value{
a list of plots. The first plot is a leaflet plot showing the locations of the points; the second plot is a time series plot showing the variation of each point’s predicted VWC overtime
}
\description{
Plotting function for the change in VWC at specific points over time.
}
\examples{
\dontrun{ plots <- site_variation(5, TRUE, project="WI_point", c("S13", "S49"))
 plots[[1]]
 plots[[2]]}
}
