% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_region_prediction.R
\name{VWC_map}
\alias{VWC_map}
\title{Regional mapping for small areas}
\usage{
VWC_map(
  area,
  start_date,
  end_date,
  resolution,
  percentile = FALSE,
  project = NULL,
  large_area = FALSE
)
}
\arguments{
\item{area}{character. Gives the name of the file that stores the study area (in shapefile format). It can be a file within the working directory or a detailed path specifying the location of the file.}

\item{start_date}{character. Defines the start date of the prediction period. The start date should be no earlier than 2016-01-01.}

\item{end_date}{character. Defines the end date of the prediction period.}

\item{resolution}{numeric. Defines the resolution of mapping region. The resolution should be smaller than or equal to 500m.}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will also be mapped.}

\item{project}{character. Defines the name of the folder to be created and save the maps. If not defined, the maps will be saved in the working directory.}
}
\value{
None
}
\description{
The major function that predicts the VWC across a region. It mostly works for small regions (preferably less than 1000 km2)
and/or low resolutions. If the downloaded maps are too large that GEE automatically splits them, the function will ask users
to consider splitting the region and mosaicking it afterwards. (direct to function: splitting())
}
\examples{
\dontrun{ VWC_map("WI.shp", "2020-06-01", "2020-09-01", 30, TRUE, "WI_region")}
}
