% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_prediction.R
\name{VWC_point}
\alias{VWC_point}
\title{Point mapping}
\usage{
VWC_point(
  file,
  start_date,
  end_date,
  percentile = FALSE,
  project = NULL,
  filename = "VWC_point_data.csv"
)
}
\arguments{
\item{file}{character. Gives the name of the file that stores the point locations (in csv format). It can be a file within the working directory or a detailed path specifying the location of the file.}

\item{start_date}{character. Defines the start date of the prediction period. The start date should be no earlier than 2016-01-01.}

\item{end_date}{character. Defines the end date of the prediction period.}

\item{percentile}{logical. If TRUE, the upper and lower bounds of the 90\% Confidence Interval of the prediction will also be mapped.}

\item{project}{character. Defines the name of the folder to be created and save the output csv. If not defined, the maps and csv file will be saved in the working directory.}

\item{filename}{character. Defines the file name where the predicted data is stored (in csv format). Default as VWC_point_data.csv.}
}
\value{
a data frame of the predicted VWC values, corresponding to the point locations.
}
\description{
A function predicting the VWC at specific points, with resolution of 30.
Rocommended: date range less than 6 months.
}
\examples{
\dontrun{ VWC_point("WI_SM.csv", "2020-06-15", "2020-09-15", TRUE, "WI_point") }
}
