% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{WI}
\alias{WI}
\title{Polygon for testing package regional mapping}
\format{
A polygon with area of 70-ha
}
\usage{
data(WI)
}
\description{
A dataset of a cropland in Rock County, WI.
Note that the data can't be directly input into the functions. To use the data,
users should first save it into a shapefile for the mapping function to access.
}
\keyword{datasets}
