% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{WI_SM}
\alias{WI_SM}
\title{Points for testing package point mapping}
\format{
12 points
}
\usage{
data(WI_SM)
}
\description{
A dataset consisting of 12 points within the cropland in Rock County, WI.
Note that the data can't be directly input into the functions. To use the data,
users should first save it into a shapefile for the mapping function to access.
}
\references{
TEROS 12, Meter Group, Inc.
}
\keyword{datasets}
